# docs/specs/openapi.yaml
openapi: 3.0.3
info:
  title: Oikonomia API
  version: 1.0.0
  description: |
    **Oikonomia** (Credit Card Analyzer) — Personal finance management platform for Argentine users.

    ## Architecture
    - **Backend**: FastAPI (Python) with PostgreSQL, Redis, and Celery
    - **Frontend**: Vite + React SPA
    - **Integrations**: Telegram Bot, WhatsApp Bot, Gemini Flash (AI), Yahoo Finance, IOL/PPI brokers

    ## Key Features
    - Expense & income tracking with encrypted fields (AES-256)
    - Credit/debit card management with billing cycle tracking
    - Hierarchical category system with auto-categorization
    - Budget management with 50/30/20 rule support
    - Investment portfolio tracking with auto price refresh
    - AI-powered spending analysis (Gemini Flash)
    - Smart CSV/PDF bank statement import
    - Family group sharing
    - Real-time SSE notifications
    - Scheduled/recurring expense automation
    - Telegram & WhatsApp bot integrations
    - Admin panel with impersonation and audit logging

    ## Security
    - JWT authentication (access + refresh tokens)
    - TOTP-based MFA
    - Field-level AES-256 encryption
    - HMAC for searchable encrypted fields
    - Rate limiting on sensitive endpoints
  contact:
    name: Oikonomia Team
  license:
    name: Proprietary
servers:
  - url: http://localhost:8001
    description: Local development
  - url: https://platform.oikonomia.ar
    description: Production
security:
  - bearerAuth: []
...
paths:
...
  /categories/suggest:
    post:
      tags: [categories]
      summary: Get LLM-based category suggestion
      operationId: suggestCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [description, amount]
              properties:
                description:
                  type: string
                amount:
                  type: number
      responses:
        '200':
          description: Suggestion result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategorySuggestResponse'
...
  /suggestions:
    get:
      tags: [suggestions]
      summary: List category suggestions
      operationId: listSuggestions
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, accepted, rejected, all]
            default: pending
      responses:
        '200':
          description: Suggestion list
          content:
            application/json:
              schema:
                type: object
                properties:
                  suggestions:
                    type: array
                    items:
                      $ref: '#/components/schemas/CategorySuggestion'
...
  /suggestions/{id}/reject:
    post:
      tags: [suggestions]
      summary: Reject a category suggestion
      operationId: rejectSuggestion
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Suggestion rejected

  # ─────────────────────────────────────────────────────────────
  # RECURRING
  # ─────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CategorySuggestResponse:
      type: object
      nullable: true
      properties:
        category_id:
          type: integer
        category_name:
          type: string
        confidence:
          type: number

    # ─── CARD CLOSINGS ────────────────────────────────────────
...
    CategorySuggestion:
      type: object
      properties:
        id:
          type: integer
        description:
          type: string
        suggested_category_id:
          type: integer
        user_id:
          type: integer
        accepted:
          type: boolean
          nullable: true

    # ─── RECURRING ────────────────────────────────────────────
...